+++ tests/cli/conftest.py
@@ -1,5 +1,7 @@
 """Pytest configuration and fixtures for coasti tests."""
 
+import shutil
+import subprocess
 import tempfile
 from pathlib import Path
 from unittest import mock
@@ -14,10 +16,10 @@ def cli_runner():
     return CliRunner()
 
 
-@pytest.fixture
-def template_bundle():
+@pytest.fixture(scope="module")
+def coasti_template_bundle():
     """
-    Create needed bundle resources.
+    Create bundle resources for the coasti template.
 
     We cannot include the bundle in git, because this would be circular, but still
     need to ship it - thus needed in tests.
@@ -38,3 +40,66 @@ def template_bundle():
             yield out_file
 
     return out_file
+
+
+@pytest.fixture(scope="class")
+def coasti_instance_dir(coasti_template_bundle):
+    """
+    A working instance of coasti with local version control.
+    """
+    from coasti import cli
+
+    cli_runner = CliRunner()
+    with tempfile.TemporaryDirectory() as tmp_path:
+        coasti_dir = Path(tmp_path) / "coasti"
+        command = ["init", "--data", '{"vcs_repo_type" : "local"}']
+        command += ["--vcs-ref", "HEAD"]
+        command += [str(coasti_dir)]
+        result = cli_runner.invoke(cli.app, command)
+        assert result.exit_code == 0, (
+            f"coasti init failed.\n"
+            f"exit_code={result.exit_code}\n"
+            f"stdout:\n{result.stdout}\n"
+            f"exception:\n{result.exception!r}\n"
+        )
+        assert coasti_dir.is_dir()
+        assert (coasti_dir / "products").is_dir()
+
+        yield coasti_dir
+
+    return coasti_dir
+
+
+@pytest.fixture(scope="class")
+def mock_product_repo():
+    """
+    Create a fake repo from which to add/install products.
+
+    We inlcude the folder structure under /templates but still need to convert
+    it into a proper git repo with tags, so that copier can install from it.
+    """
+
+    repo_root = Path(__file__).resolve().parents[2]
+    product_repo_src = repo_root / "templates" / "mock_product"
+
+    with tempfile.TemporaryDirectory() as temp_dir:
+        repo_path = Path(temp_dir) / "mock_product_repo"
+        shutil.copytree(product_repo_src, repo_path)
+
+        def run(cmd: list[str]):
+            # kwargs expansion creates type issues in subprocess.run, so we wrap it.
+            return subprocess.run(cmd, cwd=repo_path, check=True, capture_output=True)
+
+        # Initialize git repo
+        run(["git", "init"])
+        run(["git", "config", "user.email", "test@example.com"])
+        run(["git", "config", "user.name", "Test User"])
+
+        # Add and commit files
+        run(["git", "add", "."])
+        run(["git", "commit", "-m", "Initial commit"])
+
+        # Create a tag
+        run(["git", "tag", "v1.0.0"])
+
+        yield repo_path