+++ CHANGELOG.md
@@ -10,6 +10,25 @@ TLDR:
 - Features 0.1.0 (Minor Version)
 - Breaking Changes 1.0.0 (Major Version)
 
+## 0.2.0 - Upcoming
+
+### Added
+
+- We now check for coasti base dir before running product commands
+- We now check git is reachable and auth flow works before adding products
+- Product install questions no longer need to create a fake template folder
+- Better abstraction for products and their yaml writing
+- Tests for product installs with mock repo template (this can be aligned with proper demo content packages)
+- Pre-commit hook
+- `coasti product update` now takes a `--vcs-ref` (overrules and overwrites the value in products.yml)
+
+### Fixed
+
+- No longer installs on py 3.10 (which should not have been allowed) #4
+
+### Dependencies
+
+- copier 9.12.0
 
 ## 0.1.5 - 2026-02-19